Entity Update, Coin Finished
Updated the entities so that they all share the same render method -
increment the cycle, delete the old image, blit the new image. Also
finished the Coin entity, by hard coding the sprite positions in the
sprite sheet, and the number of cycles it takes.

Still need to implement the delay in sprite changes - we can put that in
Entity's definition of render()