Environment Running
Able to build the environment from the sprite sheet onto the screen
during rendering. As the image stays the same the entire game, this does
not need to dynamically change.

Environment still hasn't differentiated the tiles that the player should
not walk through. That is about two steps ahead, after we actually get
the player to move.