Coin Trajectories
Started on Coin Trajectories, tested and working in one tested case:
From left to right with no overall change in y position.

Started work on creating a variable gravity function, to create an
easing effect. So far the increase in gravity is linear, trying to
change that to make it more noticeable.