Money & Costs - Dispensation Working
Added functionality for keeping track of rising costs of money, the
money held by the player, and pretty-looking money dispensation from the
machine.

Increased the angle of trajectory for the coins, so that it looks better
when dispensed from the machine. High angle, strong force, means that
the coins might take a few seconds to land. Makes quite a nice effect of
money showering down on the player.