EntityDeleteQueue Conversion to unique_ptr
Replaced raw pointers with unique_ptrs in the EntityDeleteQueue, so that
it properly deletes the entities in the queue when cleared.

To Do: Replace the spritesheet resource pointers with shared_ptrs, so
that an object using that sprite sheet only deletes the sprite sheet if
it is the last one (after the m_spriteSheet variable in the Resources
class becomes out of scope, ie at the end of the program).