Coin Homing, SEConsole++
Adding advanced homing in coins (throwables) using a setHoming public
function that sets member variables for that throwable, determining the
maximum distance at which homing will deactivate (for the player
"outrunning" the item) and the speed at which it homes.

Upgraded the TestingConsole class to incorporate a common console
feature - the previous commands stack. Pressing UP or DOWN will traverse
a set (non-repeated items) of previous commands to help with repeatedly
called functions.