Variable Speeds
Added a new gameplay mechanic where the player gains speed from
collecting coins, and loses speed passively when not. This should
hopefully add a sense of urgency to the game, and mimic the feelings of
inability when financially weak, the sense of power in financial
abundance, and the need and panic for more money and greater values as
time goes on. Hopefully adds an element of challenge, borne from the
player's fear of inconvenience (slow player!)

Also added a few visual upgrades to the outer "wall top" tiles,
integrating them more smoothly into the environment.