Physics Refactoring
Attempted to player speed animation when above 75% speed, lead to
physics refactoring.
I wanted to do this using a particle class that extended throwable, but
didn't quite inherit from it entirely.
So I extracted a new class from Throwable: PhysicsObject
Now Throwables is a combination of PhysicsObject and Collidable.
Fully replaced the old physics code with a simpler and more easy to
understand version.

Note: Further extract physics object to modularise airborne physics
Utilize that in a more abstract way in PhysicsObject itself.