Improved Speed Particles & Dynamic Layering
Improved player speed particles to do one or any of the following:
-Fade after an instance limit is reached (ie after 100 particles are on
screen)
-Fade after a specified period
-Fade after a specified period after landing
-Fade with any speed

The angle of particles kicked up by the player now varies with their
speed - the fastest speed is indicated by more particles, at a sharper
angle.

Dynamically changes the rendering layer of the coins and particles
depending on whether or not they're airborne.