DispenseList Redesign
DispenseList has been changed from:
-the typedef of list<pair<string, int>>
to
-the typedef of list<ThrowableQuantity<Throwable>*>

Where ThrowableQuantity is a template class that uses the template
argument to construct the throwable implementation from within the
ThrowableQuantity class. Also handled the quantity of the throwable to
be thrown, which is decremented whenever a Throwable is created and
thrown from within it.

A whole mess of other changes have been made to accommodate this.